requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-ratelimit==4.1.0
django-environ==0.11.2
psycopg2-binary==2.9.9
Pillow==10.2.0
requests==2.31.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
